backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cohere==5.21.1
colorama==0.4.6
dataclasses-json==0.6.7
exceptiongroup==1.3.1
fastavro==1.12.2
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.0
idna==3.13
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-cohere==0.5.1
langchain-community==0.4.1
langchain-core==1.3.2
langchain-protocol==0.0.11
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
shellingham==1.5.4
SQLAlchemy==2.0.49
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.2
types-PyYAML==6.0.12.20260408
types-requests==2.33.0.20260408
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
